+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # BetterBlizzPlates 1.9.3
 ## Prepatch/Midnight
 ### Tweak
+- Add back support for Combo Points on Target Nameplate (Requires PRD enabled, which you can hide bars of in Edit Mode)
+- 
 - Hide absorb indicator when absorb is 0
 ### Bugfix
 - Fix Nameplate Style CVar changing on logout.